+++ js/data.js
@@ -1,0 +1,11 @@
+
+
+const types = ['palace', 'flat', 'house', 'bungalow']
+const checks = ['12:00', '13:00', '14:00']
+const titles = ['двухместный номер', 'трёхместный номер', 'номер кинг сайз']
+const rooms =  [25, 30, 45 ]
+const guests = [2, 1]
+
+export(types, checks, titles, rooms, guests)
+
+
